ADD / UPD SignUp and update role
SignUpRequest has only firstName, lastName and email now. At the signup
request we create a new user with no role, create a random password and
send a mail after the creation for giving the credentials to the new
user.
Create two new routes, one for give the president role and the other for
give the manager role. There are only a president so we remove the right
to the old.
Fix some bug on usermapper and add new exception UserAlreadyManager.